MedAgent-dev-jan/backend/requirements.txt
fastapi>=0.116.1
langchain>=0.3.27
langchain-groq
langgraph>=0.6.0
ollama>=0.5.1
pydantic>=2.11.7
python-dotenv>=1.2.1
requests>=2.32.4
streamlit>=1.47.1
twilio>=9.7.0
uvicorn>=0.35.0
